frontend toggleFavTruck impl, seperate truckCard to single file
